feat: implement mock services for isolated testing
