+++ api/src/mocks/handlers/trades.ts
@@ -1,0 +1,34 @@
+import { rest } from 'msw';
+import { mockTrades } from '../data';
+
+export const tradeHandlers = [
+  rest.get('/api/trades', (req, res, ctx) => {
+    const limit = req.url.searchParams.get('limit') || '50';
+    return res(ctx.json(mockTrades.slice(0, parseInt(limit))));
+  }),
+
+  rest.get('/api/trades/:id', (req, res, ctx) => {
+    const trade = mockTrades.find((t) => t.id === req.params.id);
+    return trade ? res(ctx.json(trade)) : res(ctx.status(404), ctx.json({ error: 'Not found' }));
+  }),
+
+  rest.post('/api/trades', (req, res, ctx) => {
+    const newTrade = { id: String(mockTrades.length + 1), ...req.body };
+    mockTrades.push(newTrade);
+    return res(ctx.status(201), ctx.json(newTrade));
+  }),
+
+  rest.patch('/api/trades/:id', (req, res, ctx) => {
+    const trade = mockTrades.find((t) => t.id === req.params.id);
+    if (!trade) return res(ctx.status(404), ctx.json({ error: 'Not found' }));
+    Object.assign(trade, req.body);
+    return res(ctx.json(trade));
+  }),
+
+  rest.delete('/api/trades/:id', (req, res, ctx) => {
+    const index = mockTrades.findIndex((t) => t.id === req.params.id);
+    if (index === -1) return res(ctx.status(404), ctx.json({ error: 'Not found' }));
+    mockTrades.splice(index, 1);
+    return res(ctx.status(204));
+  }),
+];
